add nlp.models package
