+++ sentiment_classifier/nlp/models/__init__.py
@@ -1,0 +1,1 @@
+from nlp.models.model import Model
